fix: removing word screen from screens
