+++ src/routes/Logged/index.tsx
@@ -3,11 +3,11 @@ import NavBar from '@components/NavBar'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 import Category from '@screens/Logged/Category'
 import EmailVerification from '@screens/Logged/EmailVerification'
-import HomeScreen from '@screens/Logged/HomeScreen'
-import MovieScreen from '@screens/Logged/MovieScreen'
-import PreferencesScreen from '@screens/Logged/PreferencesScreen'
-import ProfileScreen from '@screens/Logged/ProfileScreen'
-import WatchListScreen from '@screens/Logged/WatchListScreen'
+import Home from '@screens/Logged/Home'
+import Movie from '@screens/Logged/Movie'
+import Preferences from '@screens/Logged/Preferences'
+import Profile from '@screens/Logged/Profile'
+import WatchList from '@screens/Logged/WatchList'
 import { type ScreenTypes } from '@types'
 import routes from '@utils/routes'
 
@@ -16,17 +16,17 @@ const Stack = createNativeStackNavigator<ScreenTypes>()
 const tabs = {
   home: {
     label: routes.logged.home,
-    component: HomeScreen,
+    component: Home,
     icon: <Icon.Home />,
   },
   watchList: {
     label: routes.logged.watchList,
-    component: WatchListScreen,
+    component: WatchList,
     icon: <Icon.CheckCircle />,
   },
   profile: {
     label: routes.logged.profile,
-    component: ProfileScreen,
+    component: Profile,
     icon: <Icon.Person />,
   },
 }
@@ -42,12 +42,12 @@ const Logged = (
     <Stack.Screen
       options={{ gestureEnabled: false }}
       name={routes.logged.preferences}
-      component={PreferencesScreen}
+      component={Preferences}
     />
 
     <Stack.Screen
       name={routes.logged.movie}
-      component={MovieScreen}
+      component={Movie}
     />
 
     <Stack.Screen
