+++ src/screens/Logged/Home/NominationCaroussel/styles.ts
