+++ src/screens/Logged/Home/NominationItem/styles.ts
