--- src/screens/Logged/HomeScreen/index.ts
@@ -1,3 +1,0 @@
-import HomeScreen from './HomeScreen'
-
-export default HomeScreen
