+++ src/screens/Logged/Movie/index.ts
@@ -1,0 +1,3 @@
+import Movie from './Movie'
+
+export default Movie
