--- src/screens/Logged/MovieScreen/index.ts
@@ -1,3 +1,0 @@
-import MovieScreen from './MovieScreen'
-
-export default MovieScreen
