+++ src/screens/Logged/Preferences/index.ts
@@ -1,0 +1,3 @@
+import Preferences from './Preferences'
+
+export default Preferences
