+++ src/screens/Logged/Preferences/styles.ts
