--- src/screens/Logged/PreferencesScreen/index.ts
@@ -1,3 +1,0 @@
-import PreferencesScreen from './PreferencesScreen'
-
-export default PreferencesScreen
