+++ src/screens/Logged/Profile/index.ts
@@ -1,0 +1,3 @@
+import Profile from './Profile'
+
+export default Profile
