--- src/screens/Logged/ProfileScreen/index.ts
@@ -1,3 +1,0 @@
-import ProfileScreen from './ProfileScreen'
-
-export default ProfileScreen
