+++ src/screens/Logged/WatchList/index.ts
@@ -1,0 +1,3 @@
+import WatchList from './WatchList'
+
+export default WatchList
