--- src/screens/Logged/WatchListScreen/index.ts
@@ -1,3 +1,0 @@
-import WatchListScreen from './WatchListScreen'
-
-export default WatchListScreen
